+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ### Added
 
+- added unit tests to increase code coverage
 - added commands and services to send the email
 - added golang project structure
 - added basic project structure